==> 3.ipynb <==
import pandas as pd
import os
import numpy as np
import soundfile as sf
from IPython.display import clear_output
import pickle
import librosa
from sklearn.preprocessing import StandardScaler
from keras.utils import to_categorical
import tensorflow as tf
from tensorflow.keras import models, layers
from tensorflow.keras.callbacks import ReduceLROnPlateau, EarlyStopping
from tensorflow.keras.models import Model
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis as LDA

# Wczytuje embeddingi nagrań uczących model

path = '/kaggle/input/mel-train-embedding/X_train_embedding'

X_train_embedding = []

for i in range (0, 5):
    one_embedding_set = path + '_' + str(i) + '.pkl'
    
    with open(one_embedding_set, 'rb') as file:
        embedding_set = pickle.load(file)

    X_train_embedding.extend(embedding_set)

# Wczytuję targety aby stworzyć LDA
y_train = np.load('/kaggle/input/data-to-train-mel-spectrogram/y_train.npy')

# Zamiana etykiet one-hot encoded na numery klas
y_train_classes = np.argmax(y_train, axis=1)

# Przekształcenie listy embeddingów w macierz NumPy, gdzie każdy wiersz to embedding
X_train_embedding = np.vstack(X_train_embedding)

# Skalowanie danych do rozkładu o średniej 0 i odchyleniu standardowym 1
scaler = StandardScaler()
scaler.fit(X_train_embedding)
X_train_embedding = scaler.transform(X_train_embedding)

# Dopasowanie LDA (Linear Discriminant Analysis) na przeskalowanych embeddingach
lda = LDA()
lda.fit(X=X_train_embedding, y=y_train_classes);

with open('scaler_after_embedding.pkl', 'wb') as file:
    pickle.dump(scaler, file)

with open('lda.pkl', 'wb') as file:
    pickle.dump(lda, file)
